// src/modules/products/server/procedures.ts
import z from "zod";

import { baseProcedure, createTRPCRouter } from "@/trpc/init";
import type { Where } from "payload";
import { Category } from "@/payload-types";

export const productsRouter = createTRPCRouter({
    getMany: baseProcedure
    .input(
        z.object({
          category: z.string().nullable().optional(),
          minPrice: z.string().nullable().optional(),
          maxPrice: z.string().nullable().optional(),
    }),
  )
    .query(async({ ctx, input }) => {
      const where: Where = {};

      if(input.maxPrice){
        where.price = {
          ...where.price,
          less_than_equal: input.maxPrice
        }
      }

      if(input.minPrice){
        where.price = {
          ...where.price,
          greater_than_equal: input.minPrice
        }
      }

      if(input.category){
        const categoriesData = await ctx.db.find({
          collection: "categories",
          limit: 1,
          depth: 1, // Populate subcategories, subcategories.[0] will be of type "Category"
          pagination: false,
          where: {
            slug: {
              equals: input.category,
            }
          },
        });
        const formattedData = categoriesData.docs.map((doc) => ({
          ...doc,
          subcategories: (doc.subcategories?.docs ?? []).map((doc) => ({
              // because of "depth:1", confident that "doc" will be of type "Category"
              ...(doc as Category),
              subcategories:undefined,
          }))
        }));

        const subcategoriesSlugs = [];
        const parentCategory = formattedData[0];

        if(parentCategory){
          subcategoriesSlugs.push(
            ... parentCategory.subcategories.map((subcategory) => subcategory.slug)
          )
          where["category.slug"] = {
            in: [parentCategory.slug, ...subcategoriesSlugs],
          }
        }
        
      }
      const data = await ctx.db.find({
        collection: "products",
        depth: 1, // Populate "category", and "image".
        where,
      });
          
      return data;
    }),
});